+++ common/src/main/java/mtr/block/BlockPSDAPGDoorBase.java
@@ -115,7 +115,7 @@ public abstract class BlockPSDAPGDoorBase extends BlockPSDAPGBase implements Ent
 	public static abstract class TileEntityPSDAPGDoorBase extends BlockEntityClientSerializableMapper implements IGui {
 
 		private float open;
-		private float openCooldownTick;
+		private double openCooldownTick;
 
 		public TileEntityPSDAPGDoorBase(BlockEntityType<?> type, BlockPos pos, BlockState state) {
 			super(type, pos, state);
